Add the ability to remove ProtoFlux components from component search
ProtoFlux components are not useful in the generic component search (they should
be created in the world with the ProtoFlux tool), but since they are just
normal components, they appear in the search results. This change will add an
extra filter to the weighted search algorithm to downrank components that start
with the ProtoFlux path, if we're in a general component search (no search
scope).

This change can be disabled with a new config option, if for whatever reason
you want to see ProtoFlux components in the search results.